Moved data into code
HardpointActuatorApplicationSettings and
HardpointMonitorApplicationSettings data moved into code
