--- SettingFiles/Base/HardpointActuatorApplicationSettings.yaml
@@ -1,1 +1,0 @@
-HardpointActuatorTablePath: Tables/HardpointActuatorTable.csv
