--- SettingFiles/Base/HardpointMonitorApplicationSettings.yaml
@@ -1,1 +1,0 @@
-HardpointMonitorTablePath: Tables/HardpointMonitorTable.csv
