--- SettingFiles/Tables/HardpointActuatorTable.csv
@@ -1,7 +1,0 @@
-Index,ID,X Position,Y Position,Z Position,Subnet,Address,Sensor Offset,Sensor Sensitivity
-0,1,776.78278,0.00000,-89.00000,5,1,0,3
-1,2,1442.56799,0.00000,-89.00000,5,2,0,3
-2,3,2108.37793,0.00000,-89.00000,5,3,0,3
-3,4,2774.18799,0.00000,-89.00000,5,4,0,3
-4,5,3439.99805,0.00000,-89.00000,5,5,0,3
-5,6,3968.01294,0.00000,-89.00000,5,6,0,3
