--- SettingFiles/Tables/HardpointMonitorTable.csv
@@ -1,7 +1,0 @@
-Index,ID,Subnet,Address
-0,11,5,84
-1,12,5,85
-2,13,5,86
-3,14,5,87
-4,15,5,88
-5,16,5,89
