--- src/LSST/M1M3/SS/Include/HardpointMonitorTableRow.h
@@ -1,36 +1,0 @@
-/*
- * This file is part of LSST M1M3 support system package.
- *
- * Developed for the Vera C. Rubin Telescope and Site System.
- * This product includes software developed by the LSST Project
- * (https://www.lsst.org).
- * See the COPYRIGHT file at the top-level directory of this distribution
- * for details of code ownership.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-#ifndef HARDPOINTMONITORTABLEROW_H_
-#define HARDPOINTMONITORTABLEROW_H_
-
-#include <DataTypes.h>
-
-struct HardpointMonitorTableRow {
-    int32_t Index;
-    int32_t ActuatorID;
-    uint8_t Subnet;
-    uint8_t Address;
-};
-
-#endif /* HARDPOINTMONITORTABLEROW_H_ */
